+++ examples/GetFinalizedBlocks/Program.cs
@@ -16,11 +16,11 @@ internal class Program
     private static async Task GetFinalizedBlocks(ExampleOptions options)
     {
         // Construct the client.
-        using var client = new ConcordiumClient(
-            new Uri(options.Endpoint),
-            options.Port,
-            options.Timeout
-        );
+        var clientOptions = new ConcordiumClientOptions {
+            Endpoint = new Uri($"{options!.Endpoint}:{options.Port}"),
+            Timeout = TimeSpan.FromSeconds(options.Timeout)
+        };
+        using var client = new ConcordiumClient(clientOptions);
 
         // Invoke the raw call.
         var blocks = client.Raw.GetFinalizedBlocks().ResponseStream.ReadAllAsync();