+++ examples/RegisterData/Program.cs
@@ -24,11 +24,11 @@ internal class Program
         var account = WalletAccount.FromWalletKeyExportFormat(walletData);
 
         // Construct the client.
-        using var client = new ConcordiumClient(
-            new Uri(options.Endpoint),
-            options.Port,
-            options.Timeout
-        );
+        var clientOptions = new ConcordiumClientOptions {
+            Endpoint = new Uri($"{options!.Endpoint}:{options.Port}"),
+            Timeout = TimeSpan.FromSeconds(options.Timeout)
+        };
+        using var client = new ConcordiumClient(clientOptions);
 
         // Encode a string as CBOR and use that as the data to register.
         var data = OnChainData.FromTextEncodeAsCBOR(options.Data);